+++ lib/Table/TableStyle.ts
@@ -22,6 +22,7 @@ import TableTimeStyleTraits from "../Traits/TraitsClasses/TableTimeStyleTraits";
 import TableColorMap from "./TableColorMap";
 import TableColumn from "./TableColumn";
 import TableColumnType from "./TableColumnType";
+import { uniq } from "lodash-es";
 
 const DEFAULT_FINAL_DURATION_SECONDS = 3600 * 24 - 1; // one day less a second, if there is only one date.
 
@@ -328,21 +329,26 @@ export default class TableStyle {
     }
 
     const lastDate = timeColumn.valuesAsJulianDates.maximum;
-    const intervals = timeColumn.valuesAsJulianDates.values.map((date, i) => {
-      if (!date) {
-        return null;
-      }
 
-      const startDate = this.startJulianDates?.[i] ?? date;
-      const finishDate = this.finishJulianDates?.[i] ?? undefined;
+    const intervals: (TimeInterval | null)[] = new Array(
+      timeColumn.valuesAsJulianDates.values.length
+    ).fill(null);
 
-      return new TimeInterval({
+    for (let i = 0; i < timeColumn.valuesAsJulianDates.values.length; i++) {
+      const date = timeColumn.valuesAsJulianDates.values[i];
+
+      const startDate = this.startJulianDates?.[i];
+      const finishDate = this.finishJulianDates?.[i];
+
+      if (!date || !startDate || !finishDate) continue;
+
+      intervals[i] = new TimeInterval({
         start: startDate,
         stop: finishDate,
         isStopIncluded: JulianDate.equals(finishDate, lastDate),
         data: date
       });
-    });
+    }
     return intervals;
   }
 
@@ -379,23 +385,38 @@ export default class TableStyle {
 
     const firstDate = timeColumn.valuesAsJulianDates.minimum;
 
-    if (!this.timeTraits.spreadStartTime || !firstDate)
-      return timeColumn.valuesAsJulianDates.values;
+    if (!firstDate) return [];
 
-    const startDates = timeColumn.valuesAsJulianDates.values.slice();
+    // Create a new array which will be filled by rowGroups (this will exclude dates which don't have rowGroup (eg invalid regions))
+    const filteredStartDates: (JulianDate | null)[] = new Array(
+      timeColumn.valuesAsJulianDates.values.length
+    ).fill(null);
 
-    this.rowGroups.forEach(([groupId, rowIds]) => {
-      // Find row ID with earliest date in this rowGroup
-      const firstRowId = rowIds
-        .filter(id => startDates[id])
-        .sort((idA, idB) =>
-          JulianDate.compare(startDates[idA]!, startDates[idB]!)
-        )[0];
-      // Set it to earliest date in the entire column
-      if (isDefined(firstRowId)) startDates[firstRowId] = firstDate;
-    });
+    for (let i = 0; i < this.rowGroups.length; i++) {
+      const rowIds = this.rowGroups[i][1];
 
-    return startDates;
+      // Copy over valid rows
+      for (let j = 0; j < rowIds.length; j++) {
+        filteredStartDates[rowIds[j]] =
+          timeColumn.valuesAsJulianDates.values[rowIds[j]];
+      }
+
+      if (this.timeTraits.spreadStartTime) {
+        // Find row ID with earliest date in this rowGroup
+        const firstRowId = rowIds
+          .filter(id => filteredStartDates[id])
+          .sort((idA, idB) =>
+            JulianDate.compare(
+              filteredStartDates[idA]!,
+              filteredStartDates[idB]!
+            )
+          )[0];
+        // Set it to earliest date in the entire column
+        if (isDefined(firstRowId)) filteredStartDates[firstRowId] = firstDate;
+      }
+    }
+
+    return filteredStartDates;
   }
 
   /**
@@ -413,21 +434,25 @@ export default class TableStyle {
     }
 
     const startDates = timeColumn.valuesAsJulianDates.values;
+    const finishDates: (JulianDate | null)[] = new Array(
+      startDates.length
+    ).fill(null);
 
     // If displayDuration trait is set, use that to set finish date
     if (this.timeTraits.displayDuration !== undefined) {
-      return startDates.map(date =>
-        date
-          ? JulianDate.addMinutes(
-              date,
-              this.timeTraits.displayDuration!,
-              new JulianDate()
-            )
-          : null
-      );
-    }
+      for (let i = 0; i < startDates.length; i++) {
+        const date = startDates[i];
+        if (date) {
+          finishDates[i] = JulianDate.addMinutes(
+            date,
+            this.timeTraits.displayDuration!,
+            new JulianDate()
+          );
+        }
+      }
 
-    const finishDates: (JulianDate | null)[] = [];
+      return finishDates;
+    }
 
     // Otherwise estimate a final duration value to calculate the end date for groups
     // that have only one row. Fallback to a global default if an estimate
@@ -446,12 +471,26 @@ export default class TableStyle {
         startDatesForGroup,
         finalDuration
       );
-      finishDatesForGroup.forEach((date, i) => {
-        finishDates[rowIds[i]] = date;
-      });
+      for (let j = 0; j < finishDatesForGroup.length; j++) {
+        finishDates[rowIds[j]] = finishDatesForGroup[j];
+      }
     }
 
     return finishDates;
+  }
+
+  /** Returns true if rowGroups only have one unique date (i.e. they don't change over time) */
+  @computed get flatDates() {
+    if (this.rowGroups.length === 0) return false;
+
+    for (let i = 0; i < this.rowGroups.length; i++) {
+      const group = this.rowGroups[i];
+      const dates = group[1]
+        .map(row => this.timeColumn?.valuesAsDates.values[row])
+        .filter(date => date) as Date[];
+      if (uniq(dates).length > 1) return false;
+    }
+    return true;
   }
 
   /** Get rows grouped by id. Id will be calculated using idColumns, latitude/longitude columns or region column
@@ -513,9 +552,14 @@ export default class TableStyle {
         ? this.timeColumn.valuesAsJulianDates.maximum
         : sortedStartDates[sortedStartDates.length - 1];
 
-    return startDates.map(date => {
+    const finishDates: (JulianDate | null)[] = new Array(
+      startDates.length
+    ).fill(null);
+
+    for (let i = 0; i < startDates.length; i++) {
+      const date = startDates[i];
       if (!date) {
-        return null;
+        continue;
       }
 
       const nextDateIndex = binarySearch(
@@ -525,15 +569,17 @@ export default class TableStyle {
       );
       const nextDate = sortedStartDates[nextDateIndex + 1];
       if (nextDate) {
-        return nextDate;
+        finishDates[i] = nextDate;
       } else {
         // This is the last date in the row, so calculate a final date
         const finalDurationSeconds =
           estimateFinalDurationSeconds(sortedStartDates) ||
           defaultFinalDurationSeconds;
-        return addSecondsToDate(lastDate, finalDurationSeconds);
+        finishDates[i] = addSecondsToDate(lastDate, finalDurationSeconds);
       }
-    });
+    }
+
+    return finishDates;
   }
 
   private resolveColumn(name: string | undefined): TableColumn | undefined {