+++ CHANGES.md
@@ -32,6 +32,16 @@ Change Log
 * Fix `ProtomapsImageryProvider` geojson feature picking over antimeridian
 * Add Socrata group to "Add web data"
 * Fix app crash when previewing a nested reference in the catalog (eg when viewing an indexed search result where the result is a reference).
+* `TableMixin.activeStyle` will now try to pick columns which aren't labelled `"id"` or `"_id_"` by default
+* Table columns named `"easting"` and `"northing"` are now hidden by default from styles
+* Exporting `TableMixin` will now add proper file extensions
+* Added `timeDisableDimension` to `TableMixin` - this will render a checkbox to disable time dimension if `rowGroups` only have a single time interval per group (i.e. features aren't "moving" across time)
+* Added `TimeVaryingTraits.timeLabel` trait to change label on `DateTimeSelectorSection` (defaults to "Time:")
+  * This is set to `timeColumn.title`
+* `TableColumn` will try to generate prettier `title` by un-camel casing, removing underscores and capitalising words
+* Fix bug `TableColumn.type = scalar` even if there were no values.
+* `TableColumn.type = enum` requires at least 2 unique values (including null) to be selected by default
+* `TableStyle` `startDates`, `finishDates` and `timeIntervals` will only set values for valid `rowGroups` (invalid rows will be set to `null`). For example, this means that rows with invalid regions will be ignored.
 * [The next improvement]
 
 #### 8.1.4